+++ epstein_files/people/entity.py
@@ -155,7 +155,7 @@ class Entity(LoggingEntity):
         if self.is_junk:
             return self.pattern  # TODO: this sucks
 
-        return join_patterns(self._name_patterns)
+        return join_patterns(self._name_patterns + self.aliases + self.email_addresses)
 
     @property
     def identifying_strings(self) -> list[str]: